MNIST labels returned as Int8 rather than Int32

+++ examples/_common/datasets.py
@@ -2,14 +2,14 @@ import urllib.request
 from pathlib import Path
 
 import numpy as np
-from jaxtyping import Int, UInt8
+from jaxtyping import UInt8
 
 
 def load_mnist() -> tuple[
     UInt8[np.ndarray, "60000 28 28 1"],
-    Int[np.ndarray, " 60000"],
+    UInt8[np.ndarray, " 60000"],
     UInt8[np.ndarray, "10000 28 28 1"],
-    Int[np.ndarray, " 10000"],
+    UInt8[np.ndarray, " 10000"],
 ]:
     """Download MNIST dataset and return (train_images, train_labels, test_images, test_labels)."""
     cache_dir = Path.home() / ".cache" / "ion" / "mnist"
@@ -24,8 +24,11 @@ def load_mnist() -> tuple[
         )
 
     data = np.load(path)
-    train_images = data["x_train"][..., None].astype(np.uint8)
-    train_labels = data["y_train"].astype(np.int32)
-    test_images = data["x_test"][..., None].astype(np.uint8)
-    test_labels = data["y_test"].astype(np.int32)
+
+    # Add channels singleton dim to images
+    train_images = data["x_train"][..., None]
+    train_labels = data["y_train"]
+    test_images = data["x_test"][..., None]
+    test_labels = data["y_test"]
+
     return train_images, train_labels, test_images, test_labels
